Update assembly sample

+++ Sources/LibCLibrary/include/demo.h
@@ -2,4 +2,4 @@
 int demo_number();
 char* demo_string();
 double demo_compute(int n, double a, double b);
-long add_with_assembly(long a, long b);
+short add_with_assembly(short a, short b);
